# Conflicts:
#	packages/core-database/src/repositories/index.ts
#	packages/core-kernel/src/contracts/index.ts

+++ packages/core-database/src/repositories/block-repository.ts
@@ -1,10 +1,9 @@
 import { Utils } from "@arkecosystem/core-kernel";
-import { Interfaces, Managers, Transactions } from "@arkecosystem/crypto";
-import { EntityRepository, In, SelectQueryBuilder } from "typeorm";
+import { Interfaces, Transactions } from "@arkecosystem/crypto";
+import { EntityRepository, In } from "typeorm";
 
 import { Block, Round, Transaction } from "../models";
-import { AbstractEntityRepository, RepositorySearchResult } from "./repository";
-import { SearchCriteria, SearchFilter, SearchOperator, SearchPagination, SearchQueryConverter } from "./search";
+import { AbstractEntityRepository } from "./repository";
 
 @EntityRepository(Block)
 export class BlockRepository extends AbstractEntityRepository<Block> {
@@ -31,15 +30,6 @@ export class BlockRepository extends AbstractEntityRepository<Block> {
             },
             take: limit,
         });
-    }
-
-    public async findByIdOrHeight(idOrHeight: string | number): Promise<Block> {
-        try {
-            const block: Block | undefined = await this.findByHeight(idOrHeight as number);
-            return block ?? this.findById(idOrHeight as string);
-        } catch (error) {
-            return this.findById(idOrHeight as string);
-        }
     }
 
     public async findByHeight(height: number): Promise<Block | undefined> {
@@ -90,23 +80,12 @@ export class BlockRepository extends AbstractEntityRepository<Block> {
                         entity.transactions = value.map(
                             (buffer) => Transactions.TransactionFactory.fromBytesUnsafe(buffer).data,
                         );
+                    } else {
+                        entity.transactions = [];
                     }
                 },
             );
         });
-    }
-
-    public async findCommon(
-        ids: string[],
-    ): Promise<{ id: string; timestamp: number; previousBlock: string; height: string }[]> {
-        return this.createQueryBuilder()
-            .select(["id", "timestamp"])
-            .addSelect("previous_block", "previousBlock")
-            .addSelect("MAX(height)", "height")
-            .where("id IN (:...ids)", { ids })
-            .groupBy("id")
-            .orderBy("height", "DESC")
-            .getRawMany();
     }
 
     public async getStatistics(): Promise<{
@@ -182,23 +161,13 @@ export class BlockRepository extends AbstractEntityRepository<Block> {
                 });
 
                 if (block.transactions.length > 0) {
-                    let transactions = block.transactions.map((tx) =>
+                    const transactions = block.transactions.map((tx) =>
                         Object.assign(new Transaction(), {
                             ...tx.data,
                             timestamp: tx.timestamp,
                             serialized: tx.serialized,
                         }),
                     );
-
-                    // Order of transactions messed up in mainnet V1
-                    const { wrongTransactionOrder } = Managers.configManager.get("exceptions");
-                    if (wrongTransactionOrder && wrongTransactionOrder[block.data.id!]) {
-                        const fixedOrderIds = wrongTransactionOrder[block.data.id!].reverse();
-
-                        transactions = fixedOrderIds.map((id: string) =>
-                            transactions.find((transaction) => transaction.id === id),
-                        );
-                    }
 
                     transactionEntities.push(...transactions);
                 }
@@ -219,6 +188,17 @@ export class BlockRepository extends AbstractEntityRepository<Block> {
             const { round } = Utils.roundCalculator.calculateRound(lastBlockHeight);
             const blockIds = { blockIds: blocks.map((b) => b.id) };
 
+            const afterLastBlockCount = await manager
+                .createQueryBuilder()
+                .select()
+                .from(Block, "blocks")
+                .where("blocks.height > :lastBlockHeight", { lastBlockHeight })
+                .getCount();
+
+            if (afterLastBlockCount !== 0) {
+                throw new Error("Removing blocks from the middle");
+            }
+
             await manager
                 .createQueryBuilder()
                 .delete()
@@ -235,25 +215,5 @@ export class BlockRepository extends AbstractEntityRepository<Block> {
                 .where("round >= :round", { round: round + 1 })
                 .execute();
         });
-    }
-
-    public async searchByQuery(
-        query: Record<string, any>,
-        pagination: SearchPagination,
-    ): Promise<RepositorySearchResult<Block>> {
-        const filter: SearchFilter = SearchQueryConverter.toSearchFilter(query, pagination, this.metadata.columns);
-        return this.search(filter);
-    }
-
-    public async search(filter: SearchFilter): Promise<RepositorySearchResult<Block>> {
-        const queryBuilder: SelectQueryBuilder<Block> = this.createQueryBuilderFromFilter(filter);
-        const criteria: SearchCriteria[] = filter.criteria;
-
-        for (const item of criteria.filter((param) => param.operator !== SearchOperator.Custom)) {
-            const { expression, parameters } = this.criteriaToExpression(item);
-            queryBuilder.andWhere(expression, parameters);
-        }
-
-        return this.performSearch(queryBuilder);
     }
 }
